+++ website_support/models/__init__.py
@@ -1,1 +1,1 @@
-import website_support_ticket, res_partner, website_support_help
+import res_company, website_support_ticket, res_partner, website_support_help